requirements.txt
Flask
flask-sqlalchemy
flask-migrate
flask-login
Flask-WTF
Flask-Limiter
Flask-Cors
Werkzeug
psycopg2-binary
python-dotenv
gunicorn
pytest
pyotp
qrcode
requests
APScheduler
reportlab
icalendar
boto3
moto[s3]
apispec
structlog
opentelemetry-distro
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp
sentry-sdk[flask]
